+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: hnagasak <hnagasak@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/27 23:22:03 by hnagasak          #+#    #+#             */
-/*   Updated: 2024/06/27 10:15:34 by hnagasak         ###   ########.fr       */
+/*   Updated: 2024/06/27 20:09:39 by hnagasak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,16 +80,36 @@ pthread_mutex_t			**free_forks(pthread_mutex_t **forks, size_t num);
 // timer.c
 long					us2ms(long usec);
 t_timeval				us2timeval(long usec);
-void	ft_sleep(t_config conf, long msec);
+void					ft_sleep(t_config conf, long msec);
 
 // print.c
 void					print_config(t_config *config);
 void					print_philo(t_philo *data);
 void					print_forks(pthread_mutex_t **forks);
 void					print_philos_forks(t_philo *data, size_t num_of_philo);
+void					print_result(t_config *config);
+void					print_take_fork(t_philo *philo, pthread_mutex_t *fork);
 
 void					mutex_message(t_config *config, char *message);
 
-int	mutex_init(pthread_mutex_t *mutex, const pthread_mutexattr_t *attr);
+int						mutex_init(pthread_mutex_t *mutex,
+							const pthread_mutexattr_t *attr);
+
+// config
+t_config				*init_config(int argc, char **argv, t_config *config);
+void					free_config(t_config *config);
+
+// forks
+pthread_mutex_t			**init_forks(size_t num);
+void					wait_for_forks(t_philo *philo);
+void					put_forks(pthread_mutex_t *fork1,
+							pthread_mutex_t *fork2);
+
+int						should_stop(t_philo *philo);
+
+// philo
+void					*handle_philo_actions(void *args);
+void					start_philos(pthread_t *pthreads, t_philo *data,
+							size_t num_of_philo);
 
 #endif
